Разбила main на компоненты
