+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "eslint": "eslint src/**/*.js",
     "test": "npm run eslint",
-    "build": "webpack",
+    "build": "webpack"
   },
   "devDependencies": {
     "eslint": "6.1.0",
