+++ src/components/load-more.js
@@ -1,0 +1,7 @@
+const createLoadMoreButtonTemplate = () => {
+  return (
+    `<button class="load-more" type="button">load more</button>`
+  );
+};
+
+export {createLoadMoreButtonTemplate};
